Add a flag to specify whether bootnodes are also validators

+++ besu-operator/pkg/controller/besu/besu_resources.go
@@ -168,10 +168,14 @@ func (r *ReconcileBesu) besuConfigMap(instance *hyperledgerv1alpha1.Besu) *corev
 	}
 
 	GenesisObject := sampleGenesis
+	count := instance.Spec.ValidatorsCount
+	if instance.Spec.BootnodesAreValidators {
+		count = count + instance.Spec.BootnodesCount
+	}
 	GenesisObject.Blockchain = hyperledgerv1alpha1.Blockchain{
 		Nodes: hyperledgerv1alpha1.Nodes{
 			Generate: true,
-			Count:    instance.Spec.ValidatorsCount,
+			Count:    count,
 		},
 	}
 	b, err := json.Marshal(GenesisObject)
@@ -259,21 +263,34 @@ func (r *ReconcileBesu) besuInitJob(instance *hyperledgerv1alpha1.Besu) *batchv1
 								  if [ -d ${f} ]; then
 									echo $f
 									sed 's/^0x//' ${f}/key.pub > ${f}/enode.key
-									./kubectl create secret --namespace ${NAMESPACE} generic besu-validator${i}-key --from-file=private.key=${f}/key.priv --from-file=public.key=${f}/key.pub --from-file=enode.key=${f}/enode.key
+									if [ ${BOOTNODESAREVALIDATORS} = "true" ];
+									then
+										if [[ i -le ${BOOTNODES} ]]
+										then
+											./kubectl create secret --namespace ${NAMESPACE} generic besu-bootnode${i}-key --from-file=private.key=${f}/key.priv --from-file=public.key=${f}/key.pub --from-file=enode.key=${f}/enode.key
+										else
+											j=$((i - BOOTNODES))
+											./kubectl create secret --namespace ${NAMESPACE} generic besu-validator${j}-key --from-file=private.key=${f}/key.priv --from-file=public.key=${f}/key.pub --from-file=enode.key=${f}/enode.key
+										fi
+									else
+										./kubectl create secret --namespace ${NAMESPACE} generic besu-validator${i}-key --from-file=private.key=${f}/key.priv --from-file=public.key=${f}/key.pub --from-file=enode.key=${f}/enode.key
+									fi
 									i=$((i+1))
 								  fi
 								done
 
-								echo "Creating bootnode keys ..."
-								for (( j=1; j<=${BOOTNODES}; j++ ));
-								do
-									/opt/besu/bin/besu public-key export --to=public${j}.key
-									sed 's/^0x//' ./public${j}.key > enode${j}.key
-									echo "Creating bootnode ${j} secrets in k8s ..."
-									./kubectl create secret generic besu-bootnode${j}-key --namespace ${NAMESPACE} --from-file=private.key=/opt/besu/key --from-file=public.key=./public${j}.key --from-file=enode.key=./enode${j}.key
-									rm ./public${j}.key ./enode${j}.key /opt/besu/key
-								done
-			  
+								if [ ${BOOTNODESAREVALIDATORS} = "false" ];
+								then
+									echo "Creating bootnode keys ..."
+									for (( j=1; j<=${BOOTNODES}; j++ ));
+									do
+										/opt/besu/bin/besu public-key export --to=public${j}.key
+										sed 's/^0x//' ./public${j}.key > enode${j}.key
+										echo "Creating bootnode ${j} secrets in k8s ..."
+										./kubectl create secret generic besu-bootnode${j}-key --namespace ${NAMESPACE} --from-file=private.key=/opt/besu/key --from-file=public.key=./public${j}.key --from-file=enode.key=./enode${j}.key
+										rm ./public${j}.key ./enode${j}.key /opt/besu/key
+									done
+								fi
 								echo "Completed ..."`,
 							},
 							Env: []corev1.EnvVar{
@@ -284,6 +301,10 @@ func (r *ReconcileBesu) besuInitJob(instance *hyperledgerv1alpha1.Besu) *batchv1
 								{
 									Name:  "BOOTNODES",
 									Value: strconv.Itoa(instance.Spec.BootnodesCount),
+								},
+								{
+									Name:  "BOOTNODESAREVALIDATORS",
+									Value: strconv.FormatBool(instance.Spec.BootnodesAreValidators),
 								},
 							},
 						},
